Move EditPersonDescriptor to Person
