+++ src/main/java/seedu/address/model/person/Person.java
@@ -144,6 +144,44 @@ public class Person {
     }
 
     /**
+     * Stores the details to edit the person with. Each non-empty field value will replace the
+     * corresponding field value of the person.
+     */
+    public record EditPersonDescriptor(Optional<Name> name, Optional<Phone> phone, Optional<Email> email) {
+
+        // No-arg constructor, defaults all fields to Optional.empty.
+        public EditPersonDescriptor() {
+            this(Optional.empty(), Optional.empty(), Optional.empty());
+        }
+
+        // Copy constructor.
+        public EditPersonDescriptor(EditPersonDescriptor toCopy) {
+            this(toCopy.name, toCopy.phone, toCopy.email);
+        }
+
+        public EditPersonDescriptor setName(Name newName) {
+            return new EditPersonDescriptor(Optional.ofNullable(newName), this.phone, this.email);
+        }
+
+        public EditPersonDescriptor setPhone(Phone newPhone) {
+            return new EditPersonDescriptor(this.name, Optional.ofNullable(newPhone), this.email);
+        }
+
+        public EditPersonDescriptor setEmail(Email newEmail) {
+            return new EditPersonDescriptor(this.name, this.phone, Optional.ofNullable(newEmail));
+        }
+
+        @Override
+        public String toString() {
+            return new ToStringBuilder(this)
+                    .add("name", name)
+                    .add("phone", phone)
+                    .add("email", email)
+                    .toString();
+        }
+    }
+
+    /**
      * Returns true if both persons have the same identity and data fields.
      * This defines a stronger notion of equality between two persons.
      */
@@ -151,18 +189,13 @@ public class Person {
     public boolean equals(Object other) {
         if (other == this) {
             return true;
+        } else if (other instanceof Person otherPerson) {
+            return name.equals(otherPerson.name)
+                    && phone.equals(otherPerson.phone)
+                    && optionalEmail.equals(otherPerson.optionalEmail)
+                    && tags.equals(otherPerson.tags);
         }
-
-        // instanceof handles nulls
-        if (!(other instanceof Person)) {
-            return false;
-        }
-
-        Person otherPerson = (Person) other;
-        return name.equals(otherPerson.name)
-                && phone.equals(otherPerson.phone)
-                && optionalEmail.equals(otherPerson.optionalEmail)
-                && tags.equals(otherPerson.tags);
+        return false;
     }
 
     @Override
@@ -180,5 +213,4 @@ public class Person {
                 .add("tags", tags)
                 .toString();
     }
-
 }
