+++ src/test/java/seedu/address/testutil/EditPersonDescriptorBuilder.java
@@ -1,9 +1,9 @@
 package seedu.address.testutil;
 
-import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
+import seedu.address.model.person.Person.EditPersonDescriptor;
 import seedu.address.model.person.Phone;
 
 /**
